sprite system v2
instead of using image tags when declaring for sprites, I changed them to divs. Had to do some major reconstruction. Left other sprites alone due to most of them not being obvious or it doesn't really matter in their case, and sometimes even contributes to the overall order of the page/s.